// apps/frontend/src/lib/api/generated/schemas/projectResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Resume Genius API
 * OpenAPI spec version: 1.0.0
 */
import type { ProjectResponseDescription } from "./projectResponseDescription";
import type { ProjectResponseStartDate } from "./projectResponseStartDate";
import type { ProjectResponseEndDate } from "./projectResponseEndDate";
import type { ProjectResponseProjectUrl } from "./projectResponseProjectUrl";
import type { ProjectResponseRepositoryUrl } from "./projectResponseRepositoryUrl";
import type { ProjectTaskResponse } from "./projectTaskResponse";

/**
 * Response model for project.
 */
export interface ProjectResponse {
  id: string;
  user_id: string;
  project_name: string;
  description: ProjectResponseDescription;
  start_date: ProjectResponseStartDate;
  end_date: ProjectResponseEndDate;
  project_url: ProjectResponseProjectUrl;
  repository_url: ProjectResponseRepositoryUrl;
  tasks: ProjectTaskResponse[];
  created_at: string;
  updated_at: string;
}