// apps/frontend/src/lib/api/generated/schemas/projectUpdateRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Resume Genius API
 * OpenAPI spec version: 1.0.0
 */
import type { ProjectUpdateRequestProjectName } from "./projectUpdateRequestProjectName";
import type { ProjectUpdateRequestDescription } from "./projectUpdateRequestDescription";
import type { ProjectUpdateRequestStartDate } from "./projectUpdateRequestStartDate";
import type { ProjectUpdateRequestEndDate } from "./projectUpdateRequestEndDate";
import type { ProjectUpdateRequestProjectUrl } from "./projectUpdateRequestProjectUrl";
import type { ProjectUpdateRequestRepositoryUrl } from "./projectUpdateRequestRepositoryUrl";

/**
 * Request model for updating project.
 */
export interface ProjectUpdateRequest {
  project_name?: ProjectUpdateRequestProjectName;
  description?: ProjectUpdateRequestDescription;
  start_date?: ProjectUpdateRequestStartDate;
  end_date?: ProjectUpdateRequestEndDate;
  project_url?: ProjectUpdateRequestProjectUrl;
  repository_url?: ProjectUpdateRequestRepositoryUrl;
}